requirements.txt
ldap3
mysql-connector-python
